Refactored to simplify services to be shared across both platforms.
