+++ src/UI/CLI/Program.cs
@@ -1,5 +1,4 @@
 using CredBench.CLI;
-using CredBench.CLI.Services;
 using CredBench.Core.Models;
 using CredBench.Core.Services;
 using CredBench.Core.Services.CardDetectors;
